reformatting + ready for review
